pValue.jl

Documentation for pValue.jl

A set of sparse julia functions for computing p-values in a "frequentist" or "Bayesian" scenario.

pValue.jl

Functions

Index

Functions

pValue.BIC — Method

BIC(lp::AbstractFloat,ndata::Integer,nvar::Integer)

Computes the Bayes Information Criterion. 'lp' is the logarithm of the likelihood, 'ndata' the number of datapoints and 'nvar' the number of model parameters.

Examples

BIC(56.,100,3)

# output

-98.18448944203573

pValue.Frequentist_p_value — Method

Frequentist_p_value(ssrv,ndata,nvar)
Frequentist_p_value(ssrv,ndof)

Compute the 'classic' frequentist p-value. 'ssrv' is the SSR, the sum of squared residuals, 'ndata' the number of datapoints and 'nvar' the number of fit parameters. Else, 'ndof' is the number of degrees of freedom (i.e. ndata-nvar).

Examples

Frequentist_p_value(85.3,100,10)

# output

0.620453567577112

Frequentist_p_value(85.3,90)

# output

0.620453567577112

pValue.Gelman_Bayesian_p_value — Method

Gelman_Bayesian_p_value(modvecs,simvecs,obsvec,errobsvec)

Compute a 'Bayesian p-value' following the recipe by A. Gelman et al., 2013.

Explanation

'obsvec' and 'errobsvec' are length-'m' vectors of datapoints and relative uncertainties. 'modvecs' is a vector computed by the posterior distribution of parameters ('n' chains), e.g. by a MCMC, where each component is a vector of 'm' values computed using the fit function and each set of parmeters from the posterior distribution. Finally, 'simvecs' is like 'modevecs' with the addition of the predicted noise, i.e. these are simulated datapoints.

The routinely essentially compares the SSR (or, in principle, any test statistics) of each model based on the derived posterior distribution of parameters vs the data and to SSR computed by simulated data and again the posterior.

Examples

A full example of application of the GelmanBayesianp_value as well as the LucyBayesianp_value and the Frequentistpvalue is reported in this Jupyter notebook.

.

pValue.Lucy_Bayesian_p_value — Method

Lucy_Bayesian_p_value(modvecs,obsvec,errobsvec,nvars)

Compute a 'Bayesian p-value' following the recipe by L.B. Lucy, 2016, A&A 588, 19.

Explanation

'obsvec' and 'errobsvec' are length-'m' vectors of datapoints and relative uncertainties. 'modvecs' is a vector computed by the posterior distribution of parameters ('n' chains), e.g. by a MCMC, where each component is a vector of 'm' values computed using the fit function and each set of parmeters from the posterior distribution. Finally, 'nvars' is the number of parameters.

This algorithm relies on the Chi2 distribution as in the 'frequentist' case. Howver the SSR is not based only on a punt estimate but it is computed by the whole posterior distribution of parameters.

Examples

x = [1,2,3,4,5]
y = [1.01,1.95,3.05,3.97,5.1]
ey = [0.05,0.1,0.11,0.17,0.2]

f(x;a=1.,b=0.) = a.*x.+b

# Sample from a fake posterior distribution
ch = DataFrame(a=[0.99,0.95,1.01,1.02,1.03], b=[0.,-0.01,0.01,0.02,-0.01])

res = []
for i in 1:nrow(ch)
    push!(res,f(x;a=ch[i,:a],b=ch[i,:b]))
end

Lucy_Bayesian_p_value(res,y,ey,2)

# output

0.7200318895143041

pValue.RMS — Method

RMS(datavec,modvec)

Computes the Root Mean Square value. 'datavec' and 'modvec' are vectors formed by datapoints and model values to be compared with.

Examples

RMS([1.1,2.2],[1.15,2.15])

# output

0.050000000000000044

pValue.SSR — Method

SSR(modvec,obsvec,errobsvec)

Computes the Sum of Squared Residuals. 'modvec', 'obsvec' and 'errobsvec' are vectors with the model predictions, observed data and errors, respectively.

Examples

SSR([1.,2.,3.,4.],[1.1,1.9,3.05,3.8],[0.1,0.05,0.2,0.1])

# output

9.062500000000016

Index

pValue.BIC

pValue.Frequentist_p_value

pValue.Gelman_Bayesian_p_value

pValue.Lucy_Bayesian_p_value

pValue.RMS

pValue.SSR