PValue.jl

Documentation for PValue.jl

A set of sparse julia functions for computing p-values in a "frequentist" or "Bayesian" scenario.

PValue.jl

Index

Index

PValue.BIC

PValue.Frequentist_p_value

PValue.Gelman_Bayesian_p_value

PValue.GetACF

PValue.GetPACF

PValue.Lucy_Bayesian_p_value

PValue.RMS

PValue.SSR

PValue.WeightedArithmeticMean

PValue.BIC — Method

BIC(lp::AbstractFloat,ndata::Integer,nvar::Integer)

Compute the Bayes Information Criterion.

Arguments

lp logarithm of the likelihood.

ndata number of datapoints.

nvar number of model parameters.

Examples

BIC(56.,100,3)

# output

-98.18448944203573

PValue.Frequentist_p_value — Method

Frequentist_p_value(ssrv,ndata,nvar)
Frequentist_p_value(ssrv,ndof)

Compute the 'classic' frequentist p-value.

Arguments

ssrv SSR, the sum of squared residuals.

ndata number of datapoints.

nvar number of fit parameters.

ndof number of degrees of freedom

(i.e. $ndata-nvar$).

Examples

Frequentist_p_value(85.3,100,10)

# output

0.620453567577112

Frequentist_p_value(85.3,90)

# output

0.620453567577112

PValue.Gelman_Bayesian_p_value — Method

Gelman_Bayesian_p_value(modvecs,simvecs,obsvec,errobsvec)

Compute a 'Bayesian p-value' following the recipe by A. Gelman et al., 2013.

Arguments

obsvec datapoints.

errobsvec datapoint uncertainties.

modvecs model vector.

simvecs simulated vector.

Explanation

obsvec and errobsvec are length-'m' vectors of datapoints and relative uncertainties. modvecs is a vector computed by the posterior distribution of parameters (n chains), e.g. by a MCMC, where each component is a vector of m values computed using the fit function and each set of parmeters from the posterior distribution. Finally, simvecs is like modevecs with the addition of the predicted noise, i.e. these are simulated datapoints.

The routinely essentially compares the SSR (or, in principle, any test statistics) of each model based on the derived posterior distribution of parameters vs the data and to SSR computed by simulated data and again the posterior.

Examples

A full example of application of the GelmanBayesianp_value as well as the LucyBayesianp_value and the Frequentistpvalue is reported in this Jupyter notebook.

PValue.GetACF — Method

GetACF(data::Vector{Float64},lags::Integer;sigma=1.96)

Compute the [AutoCorrelation Function[(https://en.wikipedia.org/wiki/Autocorrelation) for the given lags. It returns a dictionary with the ACF and the minimum and maximum uncertainties against a white noise hypothesis.

Arguments

data logarithm of the likelihood.

lags last lag to be computed.

sigma number of sigmas for the uncertainties.

Examples

GetACF([1.2,2.5,3.5,4.3],2)["ACF"]

# output

3-element Vector{Float64}:
  1.0
  0.23928737773637632
 -0.2945971122496507

PValue.GetPACF — Method

GetPACF(data::Vector{Float64},lags::Integer;sigma=1.96)

Compute the [Partial AutoCorrelation Function[(https://en.wikipedia.org/wiki/Partialautocorrelationfunction) for the given lags. It returns a dictionary with the PACF and the minimum and maximum uncertainties.

Arguments

data logarithm of the likelihood.

lags last lag to be computed.

sigma number of sigmas for the uncertainties.

Examples

GetPACF([1.2,2.5,3.5,4.3],1)["PACF"]

# output

2-element Vector{Float64}:
 1.0
 0.7819548872180438

PValue.Lucy_Bayesian_p_value — Method

Lucy_Bayesian_p_value(modvecs,obsvec,errobsvec,nvars)

Compute a 'Bayesian p-value' following the recipe by L.B. Lucy, 2016, A&A 588, 19.

Arguments

obsvec datapoints.

errobsvec datapoint uncertainties.

nvars number of parameters.

Explanation

obsvec and errobsvec are length-m vectors of datapoints and relative uncertainties. modvecs is a vector computed by the posterior distribution of parameters (n chains), e.g. by a MCMC, where each component is a vector of m values computed using the fit function and each set of parmeters from the posterior distribution. Finally, nvars is the number of parameters.

This algorithm relies on the Chi2 distribution as in the 'frequentist' case. Howver the SSR is not based only on a punt estimate but it is computed by the whole posterior distribution of parameters.

Examples

x = [1,2,3,4,5]
y = [1.01,1.95,3.05,3.97,5.1]
ey = [0.05,0.1,0.11,0.17,0.2]

f(x;a=1.,b=0.) = a.*x.+b

# Sample from a fake posterior distribution
ch = DataFrame(a=[0.99,0.95,1.01,1.02,1.03], b=[0.,-0.01,0.01,0.02,-0.01])

res = []
for i in 1:nrow(ch)
    push!(res,f(x;a=ch[i,:a],b=ch[i,:b]))
end

Lucy_Bayesian_p_value(res,y,ey,2)

# output

0.7200318895143041

PValue.RMS — Method

RMS(datavec,modvec)

Compute the Root Mean Square value.

Arguments

datavec datapoints.

modvec model values.

Examples

RMS([1.1,2.2],[1.15,2.15])

# output

0.050000000000000044

PValue.SSR — Method

SSR(modvec,obsvec,errobsvec)

Compute the Sum of Squared Residuals.

Arguments

modvec model predictions.

obsvec observed data.

`errobsvec~ uncertainties.

Examples

SSR([1.,2.,3.,4.],[1.1,1.9,3.05,3.8],[0.1,0.05,0.2,0.1])

# output

9.062500000000016

PValue.WeightedArithmeticMean — Method

WeightedArithmeticMean(x,ex)

Compute the Weighted Arithmetic Mean.

Arguments

x input vector

ex uncertainties.

Examples

x = [1.2,2.2,4.5,3,3.6]
ex = [0.2,0.2,0.5,0.1,0.6]

WeightedArithmeticMean(x,ex)

# output

(2.634301913536499, 0.07986523020975032)